% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeExists.R
\name{attributeExists}
\alias{attributeExists}
\title{Check if an Attribute exists}
\usage{
attributeExists(handle, name)
}
\arguments{
\item{handle}{External pointer to a HDF5 Group or DataSet.}

\item{name}{String containing the name of the Attribute.}
}
\value{
Boolean indicating whether the named Attribute exists.
}
\description{
Check if an Attribute exists on a HDF5 Group or DataSet.
}
\examples{
tmp <- tempfile(fileext = ".h5")
fhandle <- createFile(tmp)
ghandle <- createGroup(fhandle, "foo")

createAttribute(
    ghandle,
    "bar",
    type = createIntegerType(bits = 32, sign = TRUE),
    dims = NULL
)

createAttribute(
    ghandle,
    "whee",
    type = createStringType(length = 10)
    dims = NULL
)

attributeExists(ghandle, "bar")
attributeExists(ghandle, "whee")
attributeExists(ghandle, "stuff")

closeGroup(ghandle)
closeGroup(fhandle)

}
\author{
Aaron Lun
}
