% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chihaya-package.R
\docType{package}
\name{chihaya-package}
\alias{chihaya}
\alias{chihaya-package}
\title{chihaya: HDF5 Bindings for Alabaster}
\description{
Low-level utilities for interacting with HDF5 files via C++ bindings to the Rhdf5lib package. This is intended to be used by the various alabaster.* packages, where it provides more control than the rhdf5 package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ArtifactDB/chihaya-R}
  \item Report bugs at \url{https://github.com/ArtifactDB/chihaya-R/issues}
}

}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}

Authors:
\itemize{
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}
}

}
