% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childExists.R
\name{childExists}
\alias{childExists}
\title{Check if a child object exists}
\usage{
childExists(handle, name)
}
\arguments{
\item{handle}{External pointer to a HDF5 Group.}

\item{name}{String containing the name of the child object.}
}
\value{
Boolean indicating whether the named child object exists.
}
\description{
Check if a child object exists in a HDF5 Group.
}
\examples{
tmp <- tempfile(fileext=".h5")
fhandle <- createFile(tmp)
ghandle <- createGroup(fhandle, "foo")

attributeExists(fhandle, "foo")
attributeExists(fhandle, "bar")

closeGroup(ghandle)
closeGroup(fhandle)

}
\author{
Aaron Lun
}
