% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAttribute.R
\name{createAttribute}
\alias{createAttribute}
\title{Create a HDF5 Attribute}
\usage{
createAttribute(handle, name, type, dims)
}
\arguments{
\item{handle}{External pointer to a HDF5 Group or DataSet.}

\item{name}{String containing the name of the Attribute.}

\item{type}{List specifying the HDF5 type of the new Attribute,
see \code{\link{createStringType}} and related functions.}

\item{dims}{Numeric vector of dimension extents.
An empty vector or \code{NULL} indicates that the new Attribute is a scalar.}
}
\value{
External pointer to a HDF5 Attribute.
}
\description{
Create an Attribute on a HDF5 Group or DataSet.
}
\examples{
tmp <- tempfile(fileext = ".h5")
fhandle <- createFile(tmp)
ghandle <- createGroup(fhandle, "foo")

createAttribute(
    ghandle,
    "bar",
    type = createIntegerType(bits = 32, sign = TRUE),
    dims = NULL
)

createAttribute(
    ghandle,
    "whee",
    type = createStringType(length = 10),
    dims = c(4, 3)
)

listAttributes(ghandle)
closeGroup(ghandle)
closeGroup(fhandle)

}
\author{
Aaron Lun
}
