% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataSet.R
\name{createDataSet}
\alias{createDataSet}
\title{Create a HDF5 DataSet}
\usage{
createDataSet(handle, name, type, dims, compress = 6, chunkdims = NULL)
}
\arguments{
\item{handle}{External pointer to a HDF5 Group.}

\item{name}{String containing the name of the DataSet.}

\item{type}{List specifying the HDF5 type of the new DataSet,
see \code{\link{createStringType}} and related functions.}

\item{dims}{Numeric vector of dimension extents.
An empty vector or \code{NULL} indicates that the new attribute is a scalar.}

\item{compress}{Integer specifying the DEFLATE compression level.
Larger values increase compression efficiency at the cost of extra compute.
If set to zero, no compression is performed.}

\item{chunkdims}{Numeric vector of chunk dimension extents.
This should be of the same length as \code{dims}.
Ignored if \code{compress = 0} or any entry of \code{dims} is zero.}
}
\value{
External pointer to a HDF5 DataSet.
}
\description{
Create a new DataSet in a HDF5 Group.
}
\examples{
tmp <- tempfile(fileext = ".h5")
fhandle <- createFile(tmp)
createDataSet(
    ghandle,
    "bar",
    type = createIntegerType(bits = 32, sign = TRUE),
    dims = c(10, 20, 5) 
)
closeGroup(fhandle)

}
\author{
Aaron Lun
}
