% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFile.R
\name{createFile}
\alias{createFile}
\title{Create a HDF5 file}
\usage{
createFile(path)
}
\arguments{
\item{path}{String containing the path to a new HDF5 file.
If any existing file is present, it will be overwritten.}
}
\value{
An external pointer to a writeable HDF5 group representing the new file.
}
\description{
Create a new HDF5 file.
}
\examples{
tmp <- tempfile(fileext=".h5")
handle <- createFile(tmp)
closeGroup(handle)
file.info(tmp)

}
\author{
Aaron Lun
}
