% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGroup.R
\name{createGroup}
\alias{createGroup}
\title{Create a HDF5 Group}
\usage{
createGroup(handle, name)
}
\arguments{
\item{handle}{External pointer to a HDF5 Group.}

\item{name}{String containing the name of the Group.}
}
\value{
External pointer to a HDF5 Group.
}
\description{
Create a new HDF5 Group.
}
\examples{
tmp <- tempfile(fileext = ".h5")
fhandle <- createFile(tmp)
ghandle <- createGroup(fhandle, "foo")
ghandle2 <- createGroup(ghandle, "bar")
closeGroup(ghandle2)
closeGroup(ghandle)
closeGroup(fhandle)

}
\author{
Aaron Lun
}
