% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createType.R
\name{createType}
\alias{createType}
\alias{createStringType}
\alias{createIntegerType}
\alias{createFloatType}
\title{Specify a HDF5 data type}
\usage{
createStringType(length, encoding = c("UTF-8", "ASCII"))

createIntegerType(bits, sign)

createFloatType(bits)
}
\arguments{
\item{length}{Positive integer specifying the maximum length of a fixed-length string.
Alternatively zero or \code{NULL}, in which case a variable-length string type is created.}

\item{encoding}{String specifying the encoding of the string.}

\item{bits}{Integer specifying the number of bits in the integer or floating-point number.
For \code{createIntegerType}, this should be one of 8, 16, 32 or 64.
For \code{createFloatType}, this should be one of 32 or 64.}

\item{sign}{Boolean specifying whether the integer is signed.}
}
\value{
List containing information about the HDF5 data type.
}
\description{
Create a specification for a HDF5 data type,
to be used in functions like \code{\link{createDataSet}} and \code{\link{createAttribute}}.
}
\examples{
createStringType(10)
createIntegerType(8, TRUE)
createFloatType(64)

}
\author{
Aaron Lun
}
