% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAttributes.R
\name{listAttributes}
\alias{listAttributes}
\title{List attributes}
\usage{
listAttributes(handle)
}
\arguments{
\item{handle}{External pointer to a HDF5 Group or DataSet.}
}
\value{
Character vector of attribute names.
}
\description{
List attributes on a HDF5 Group or DataSet.
}
\examples{
tmp <- tempfile(fileext=".h5")
fhandle <- createFile(tmp)
ghandle <- createGroup(fhandle, "foo")

createAttribute(
    ghandle,
    "bar",
    type=createIntegerType(bits=32, sign=TRUE)
)

createAttribute(
    ghandle,
    "whee",
    type=createStringType(length=10)
)

listAttributes(ghandle)
closeGroup(ghandle)
closeGroup(fhandle)

}
\author{
Aaron Lun
}
