% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listChildren.R
\name{listChildren}
\alias{listChildren}
\title{List child objects}
\usage{
listChildren(handle)
}
\arguments{
\item{handle}{External pointer to a HDF5 Group.}
}
\value{
List of two character vectors \code{group} and \code{dataset},
containing the names of the child Groups and DataSets of the HDF5 Group referenced by \code{handle}.
}
\description{
List child objects of a HDF5 Group.
}
\examples{
tmp <- tempfile(fileext=".h5")
fhandle <- createFile(tmp)
ghandle <- createGroup(fhandle, "foo")
dhandle <- createDataSet(
    fhandle,
    "bar",
    type = createFloatType(32),
    dims = NULL
)

listChildren(fhandle)
closeGroup(ghandle)
closeDataSet(dhandle)
closeGroup(fhandle)

}
\author{
Aaron Lun
}
