% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openAttribute.R
\name{openAttribute}
\alias{openAttribute}
\title{Open and close a HDF5 Attribute}
\usage{
openAttribute(handle, name)
}
\arguments{
\item{handle}{For \code{openAttribute}, an external pointer to a HDF5 Group or DataSet.
This can returned by \code{\link{openGroup}} or other functions like \code{\link{createDataSet}}.

For \code{closeAttribute}, an external pointer to a HDF5 Attribute.
This can returned by \code{\link{openAttribute}} or other functions like \code{\link{createAttribute}}.}

\item{name}{String containing the name of the child Attribute to open.}
}
\value{
For \code{openAttribute}, an external pointer to the HDF5 Attribute. 

For \code{closeGroup}, the Attribute referenced by \code{handle} is closed.
\code{handle} itself should no longer be used.
\code{NULL} is invisibly returned.
}
\description{
Open and close an existing HDF5 Attribute.
}
\details{
It is not strictly necessary to call \code{closeAttribute} as the Attribute will be automatically closed upon garbage collection of the relevant pointers.
Nonetheless, it can be helpful to explicitly close the group to release system resources (e.g., memory, file handles) in a predictable manner.
}
\examples{
tmp <- tempfile(fileext=".h5")
fhandle <- createFile(tmp)
ghandle <- createGroup(fhandle, "foo")
ahandle <- createAttribute(
    ghandle,
    "bar",
    type = createFloatType(64),
    dim = NULL
)
closeAttribute(ahandle)
closeGroup(ghandle)
closeGroup(fhandle)

fhandle <- openFile(handle)
ghandle <- openGroup(fhandle, "foo")
ahandle <- openAttribute(ghandle, "bar")
closeAttribute(ahandle)
closeGroup(ghandle)
closeGroup(fhandle)

}
\author{
Aaron Lun
}
