% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFile.R
\name{openFile}
\alias{openFile}
\title{Open a HDF5 file}
\usage{
openFile(path, read.only = TRUE)
}
\arguments{
\item{path}{String containing the path to an existing HDF5 file.}

\item{read.only}{Boolean indicating whether to open the file in read-only mode.
If \code{FALSE}, the file is opened in read-write mode.}
}
\value{
External pointer to a HDF5 group representing the file contents.
}
\description{
Open an existing HDF5 file.
}
\examples{
tmp <- tempfile(fileext=".h5")
handle <- createFile(tmp)
closeGroup(handle)

new_handle <- openFile(tmp)
listChildren(new_handle)
closeGroup(new_handle)

}
\author{
Aaron Lun
}
