% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAttribute.R
\name{writeAttribute}
\alias{writeAttribute}
\title{Write a HDF5 Attribute}
\usage{
writeAttribute(handle, value)
}
\arguments{
\item{handle}{External pointer to a HDF5 Attribute,
typically returned by \code{\link{openAttribute}} or \code{\link{createAttribute}}.}

\item{value}{Value of the HDF5 attribute as an atomic vector.}
}
\value{
\code{value} is written to \code{handle}, and \code{NULL} is invisibly returned.
}
\description{
Write data to an existing HDF5 Attribute.
}
\examples{
tmp <- tempfile(fileext=".h5")
fhandle <- createFile(tmp)
ghandle <- createGroup(fhandle, "foo")
ahandle <- createAttribute(
    ghandle,
    "bar",
    type = createFloatType(64),
    dim = NULL
)
writeAttribute(ahandle, 123.0)
closeAttribute(ahandle)
closeGroup(ghandle)
closeGroup(fhandle)

}
\author{
Aaron Lun
}
