% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDataSet.R
\name{writeDataSet}
\alias{writeDataSet}
\title{Write a HDF5 DataSet}
\usage{
writeDataSet(handle, value)
}
\arguments{
\item{handle}{External pointer to a HDF5 DataSet,
typically returned by \code{\link{openDataSet}} or \code{\link{createDataSet}}.}

\item{value}{Value of the HDF5 attribute as an atomic vector.}
}
\value{
\code{value} is written to \code{handle}, and \code{NULL} is invisibly returned.
}
\description{
Write data to an existing HDF5 DataSet.
}
\examples{
tmp <- tempfile(fileext=".h5")
fhandle <- createFile(tmp)
dhandle <- createDataSet(
    fhandle,
    "bar",
    type = createFloatType(64),
    dim = c(10, 20)
)
writeDataSet(dhandle, 1:200)
closeDataSet(dhandle)
closeGroup(fhandle)

}
\author{
Aaron Lun
}
