% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openGroup.R
\name{openGroup}
\alias{openGroup}
\title{Open and close a HDF5 Group}
\usage{
openGroup(handle, name)
}
\arguments{
\item{handle}{External pointer to a HDF5 Group.
This can returned by \code{openGroup} or other functions like \code{\link{openFile}} or \code{\link{createFile}}.}

\item{name}{String containing the name of the child Group to open.}
}
\value{
For \code{openGroup}, an external pointer to the HDF5 group. 

For \code{closeGroup}, the Group referenced by \code{handle} is closed.
\code{handle} itself should no longer be used.
\code{NULL} is invisibly returned.
}
\description{
Open and close an existing HDF5 Group.
}
\details{
It is not strictly necessary to call \code{closeGroup} as the Group will be automatically closed upon garbage collection of the relevant pointers.
Nonetheless, it can be helpful to explicitly close the group to release system resources (e.g., memory, file handles) in a predictable manner.
}
\examples{
tmp <- tempfile(fileext=".h5")
handle <- createFile(tmp)
createGroup(handle, "foo")
closeFile(handle)

new_handle <- openFile(handle)
ghandle <- openGroup(new_handle, "foo")
closeGroup(ghandle)
closeGroup(new_handle)

}
\author{
Aaron Lun
}
